refactor(executor): Split executor.rs into logical execution modules
Refactors the 1,386-line columnar/executor.rs into separate modules
for better code organization:

- executor/mod.rs (322 lines): Main entry point and orchestration
- executor/fused.rs (313 lines): Fused filter+aggregate optimization
- executor/aggregate.rs (393 lines): Standard aggregation functions
- executor/expression.rs (341 lines): Expression evaluation helpers

All modules are now under 400 lines for easier maintenance.
No functional changes - all existing tests pass.

Closes #3106
